pyproject.toml
[project]
name = "StarResonanceFishing"
dynamic = ["version"]
description = "A bot for fishing in Star Resonance."
requires-python = ">=3.14"

dependencies = [
    "numpy>=2.4.1",
    "opencv-python>=4.8.1.78",
    "pyautogui>=0.9.54",
    "pygetwindow>=0.0.9",
    "pytesseract>=0.3.13",
    "pyyaml>=6.0.1",
]

[dependency-groups]
dev = [
    "pip>=25.3",
    "pyinstaller>=6.18.0",
    "ruff>=0.14.13",
    "ty>=0.0.13",
]

[tool.ruff]
line-length = 79
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "UP"]
